Refactor: Move screen components to a dedicated `component` package
This commit refactors the project structure by moving several UI components from the `screen` package to a more appropriate `component` package within the `feature/home` module. This improves organization by separating reusable components from top-level screen composables.

The affected components are related to item resizing and popups, which are used across different screens like the home pager, folder view, and resize overlays.

### Key Changes:

*   **Component Relocation:**
    *   `GridItemResizeOverlay.kt` and `WidgetGridItemResizeOverlay.kt` have been moved from `.../screen/resize` to `.../component/resize`.
    *   `ResizeOverlay.kt` has been extracted from `ResizeScreen.kt` and moved to `.../component/resize`.
    *   `GridItemPopup.kt` and its related content have been extracted from `.../screen/pager/Popup.kt` and moved to a new file at `.../component/popup/GridItemPopup.kt`.

*   **Package and Import Updates:**
    *   The package declarations in the moved files have been updated to reflect their new location (e.g., `...screen.resize` to `...component.resize`).
    *   Imports in `HomeScreen.kt`, `HorizontalPagerScreen.kt`, `FolderScreen.kt`, `FolderResizeScreen.kt`, and `ResizeScreen.kt` have been updated to point to the new component locations.

*   **Code Cleanup:**
    *   The now-empty `ResizeOverlay` composable and related code have been removed from `ResizeScreen.kt`.
    *   The `GridItemPopup` composable and its helper functions have been removed from `.../screen/pager/Popup.kt`, which now only contains `SettingsPopup`.
    *   A minor addition was made in `HomeScreen.kt` to handle page count for the new `Screen.FolderResize` state.
