+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/component/popup/GridItemPopup.kt
@@ -1,0 +1,469 @@
+/*
+ *
+ *   Copyright 2023 Einstein Blanco
+ *
+ *   Licensed under the GNU General Public License v3.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       https://www.gnu.org/licenses/gpl-3.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+package com.eblan.launcher.feature.home.component.popup
+
+import android.appwidget.AppWidgetProviderInfo
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.calculateStartPadding
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.Layout
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.unit.LayoutDirection
+import androidx.compose.ui.unit.dp
+import com.eblan.launcher.designsystem.icon.EblanLauncherIcons
+import com.eblan.launcher.domain.model.EblanAppWidgetProviderInfo
+import com.eblan.launcher.domain.model.EblanApplicationInfoGroup
+import com.eblan.launcher.domain.model.EblanShortcutInfo
+import com.eblan.launcher.domain.model.EblanShortcutInfoByGroup
+import com.eblan.launcher.domain.model.GridItem
+import com.eblan.launcher.domain.model.GridItemData
+import com.eblan.launcher.domain.model.GridItemSettings
+import com.eblan.launcher.feature.home.model.Drag
+import com.eblan.launcher.feature.home.model.GridItemSource
+import com.eblan.launcher.feature.home.model.SharedElementKey
+
+@Composable
+internal fun GridItemPopup(
+    modifier: Modifier = Modifier,
+    gridItem: GridItem,
+    popupIntOffset: IntOffset,
+    popupIntSize: IntSize,
+    eblanShortcutInfosGroup: Map<EblanShortcutInfoByGroup, List<EblanShortcutInfo>>,
+    hasShortcutHostPermission: Boolean,
+    currentPage: Int,
+    drag: Drag,
+    gridItemSettings: GridItemSettings,
+    eblanAppWidgetProviderInfosGroup: Map<String, List<EblanAppWidgetProviderInfo>>,
+    paddingValues: PaddingValues,
+    onEdit: (String) -> Unit,
+    onResize: () -> Unit,
+    onWidgets: (EblanApplicationInfoGroup) -> Unit,
+    onDeleteGridItem: (GridItem) -> Unit,
+    onInfo: (Long, String) -> Unit,
+    onDismissRequest: () -> Unit,
+    onTapShortcutInfo: (Long, String, String) -> Unit,
+    onLongPressGridItem: (
+        gridItemSource: GridItemSource,
+        imageBitmap: ImageBitmap?,
+    ) -> Unit,
+    onUpdateGridItemOffset: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onDraggingGridItem: () -> Unit,
+    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
+) {
+    val density = LocalDensity.current
+
+    val leftPadding = with(density) {
+        paddingValues.calculateStartPadding(LayoutDirection.Ltr).roundToPx()
+    }
+
+    val topPadding = with(density) {
+        paddingValues.calculateTopPadding().roundToPx()
+    }
+    val x = popupIntOffset.x - leftPadding
+
+    val y = popupIntOffset.y - topPadding
+
+    Layout(
+        modifier = modifier
+            .pointerInput(Unit) {
+                detectTapGestures(
+                    onPress = {
+                        awaitRelease()
+
+                        onDismissRequest()
+                    },
+                )
+            }
+            .fillMaxSize()
+            .padding(paddingValues),
+        content = {
+            GridItemPopupContent(
+                modifier = Modifier.padding(5.dp),
+                eblanShortcutInfosGroup = eblanShortcutInfosGroup,
+                gridItem = gridItem,
+                hasShortcutHostPermission = hasShortcutHostPermission,
+                currentPage = currentPage,
+                drag = drag,
+                gridItemSettings = gridItemSettings,
+                eblanAppWidgetProviderInfosGroup = eblanAppWidgetProviderInfosGroup,
+                onEdit = onEdit,
+                onDismissRequest = onDismissRequest,
+                onResize = onResize,
+                onInfo = onInfo,
+                onWidgets = onWidgets,
+                onDeleteGridItem = onDeleteGridItem,
+                onTapShortcutInfo = onTapShortcutInfo,
+                onLongPressGridItem = onLongPressGridItem,
+                onUpdateGridItemOffset = onUpdateGridItemOffset,
+                onDraggingGridItem = onDraggingGridItem,
+                onUpdateSharedElementKey = onUpdateSharedElementKey,
+            )
+        },
+    ) { measurables, constraints ->
+        val placeable = measurables.first().measure(
+            constraints.copy(
+                minWidth = 0,
+                minHeight = 0,
+            ),
+        )
+
+        val parentCenterX = x + popupIntSize.width / 2
+
+        val childX = (parentCenterX - placeable.width / 2)
+            .coerceIn(0, constraints.maxWidth - placeable.width)
+
+        val topY = y - placeable.height
+        val bottomY = y + popupIntSize.height
+
+        val childY = if (topY < 0) bottomY else topY
+
+        layout(constraints.maxWidth, constraints.maxHeight) {
+            placeable.place(
+                x = childX,
+                y = childY,
+            )
+        }
+    }
+}
+
+@Composable
+private fun GridItemPopupContent(
+    modifier: Modifier = Modifier,
+    eblanShortcutInfosGroup: Map<EblanShortcutInfoByGroup, List<EblanShortcutInfo>>,
+    gridItem: GridItem,
+    hasShortcutHostPermission: Boolean,
+    currentPage: Int,
+    drag: Drag,
+    gridItemSettings: GridItemSettings,
+    eblanAppWidgetProviderInfosGroup: Map<String, List<EblanAppWidgetProviderInfo>>,
+    onEdit: (String) -> Unit,
+    onDismissRequest: () -> Unit,
+    onResize: () -> Unit,
+    onInfo: (
+        serialNumber: Long,
+        componentName: String,
+    ) -> Unit,
+    onWidgets: (EblanApplicationInfoGroup) -> Unit,
+    onDeleteGridItem: (GridItem) -> Unit,
+    onTapShortcutInfo: (
+        serialNumber: Long,
+        packageName: String,
+        shortcutId: String,
+    ) -> Unit,
+    onLongPressGridItem: (
+        gridItemSource: GridItemSource,
+        imageBitmap: ImageBitmap?,
+    ) -> Unit,
+    onUpdateGridItemOffset: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onDraggingGridItem: () -> Unit,
+    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
+) {
+    Surface(
+        modifier = modifier,
+        shape = RoundedCornerShape(30.dp),
+        shadowElevation = 2.dp,
+        content = {
+            when (val data = gridItem.data) {
+                is GridItemData.ApplicationInfo -> {
+                    ApplicationInfoGridItemMenu(
+                        modifier = modifier,
+                        eblanShortcutInfosByPackageName = eblanShortcutInfosGroup[
+                            EblanShortcutInfoByGroup(
+                                serialNumber = data.serialNumber,
+                                packageName = data.packageName,
+                            ),
+                        ],
+                        hasShortcutHostPermission = hasShortcutHostPermission,
+                        currentPage = currentPage,
+                        drag = drag,
+                        icon = data.icon,
+                        gridItemSettings = gridItemSettings,
+                        eblanAppWidgetProviderInfosByPackageName = eblanAppWidgetProviderInfosGroup[data.packageName],
+                        onEdit = {
+                            onDismissRequest()
+
+                            onEdit(gridItem.id)
+                        },
+                        onResize = {
+                            onResize()
+
+                            onDismissRequest()
+                        },
+                        onInfo = {
+                            onInfo(
+                                data.serialNumber,
+                                data.componentName,
+                            )
+
+                            onDismissRequest()
+                        },
+                        onDelete = {
+                            onDeleteGridItem(gridItem)
+
+                            onDismissRequest()
+                        },
+                        onWidgets = {
+                            onWidgets(
+                                EblanApplicationInfoGroup(
+                                    serialNumber = data.serialNumber,
+                                    packageName = data.packageName,
+                                    icon = data.icon,
+                                    label = data.label,
+                                ),
+                            )
+
+                            onDismissRequest()
+                        },
+                        onTapShortcutInfo = { serialNumber, packageName, shortcutId ->
+                            onTapShortcutInfo(
+                                serialNumber,
+                                packageName,
+                                shortcutId,
+                            )
+
+                            onDismissRequest()
+                        },
+                        onLongPressGridItem = onLongPressGridItem,
+                        onUpdateGridItemOffset = onUpdateGridItemOffset,
+                        onDraggingGridItem = {
+                            onDraggingGridItem()
+
+                            onDismissRequest()
+                        },
+                        onUpdateSharedElementKey = onUpdateSharedElementKey,
+                    )
+                }
+
+                is GridItemData.Folder, is GridItemData.ShortcutInfo, is GridItemData.ShortcutConfig -> {
+                    GridItemMenu(
+                        modifier = modifier,
+                        onEdit = {
+                            onEdit(gridItem.id)
+
+                            onDismissRequest()
+                        },
+                        onResize = {
+                            onResize()
+
+                            onDismissRequest()
+                        },
+                        onDelete = {
+                            onDeleteGridItem(gridItem)
+
+                            onDismissRequest()
+                        },
+                    )
+                }
+
+                is GridItemData.Widget -> {
+                    val showResize = data.resizeMode != AppWidgetProviderInfo.RESIZE_NONE
+
+                    WidgetGridItemMenu(
+                        modifier = modifier,
+                        showResize = showResize,
+                        onResize = {
+                            onResize()
+
+                            onDismissRequest()
+                        },
+                        onDelete = {
+                            onDeleteGridItem(gridItem)
+
+                            onDismissRequest()
+                        },
+                    )
+                }
+            }
+        },
+    )
+}
+
+@Composable
+private fun ApplicationInfoGridItemMenu(
+    modifier: Modifier = Modifier,
+    eblanShortcutInfosByPackageName: List<EblanShortcutInfo>?,
+    hasShortcutHostPermission: Boolean,
+    currentPage: Int,
+    drag: Drag,
+    icon: String?,
+    gridItemSettings: GridItemSettings,
+    eblanAppWidgetProviderInfosByPackageName: List<EblanAppWidgetProviderInfo>?,
+    onEdit: () -> Unit,
+    onResize: () -> Unit,
+    onInfo: () -> Unit,
+    onDelete: () -> Unit,
+    onWidgets: () -> Unit,
+    onTapShortcutInfo: (
+        serialNumber: Long,
+        packageName: String,
+        shortcutId: String,
+    ) -> Unit,
+    onLongPressGridItem: (
+        gridItemSource: GridItemSource,
+        imageBitmap: ImageBitmap?,
+    ) -> Unit,
+    onUpdateGridItemOffset: (
+        intOffset: IntOffset,
+        intSize: IntSize,
+    ) -> Unit,
+    onDraggingGridItem: () -> Unit,
+    onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
+) {
+    Column(
+        modifier = modifier,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        if (hasShortcutHostPermission &&
+            !eblanShortcutInfosByPackageName.isNullOrEmpty()
+        ) {
+            ShortcutInfoMenu(
+                modifier = modifier,
+                currentPage = currentPage,
+                drag = drag,
+                icon = icon,
+                eblanShortcutInfosGroup = eblanShortcutInfosByPackageName,
+                gridItemSettings = gridItemSettings,
+                onTapShortcutInfo = onTapShortcutInfo,
+                onLongPressGridItem = onLongPressGridItem,
+                onUpdateGridItemOffset = onUpdateGridItemOffset,
+                onDraggingGridItem = onDraggingGridItem,
+                onUpdateSharedElementKey = onUpdateSharedElementKey,
+            )
+
+            Spacer(modifier = Modifier.height(5.dp))
+        }
+
+        Row {
+            IconButton(
+                onClick = onEdit,
+            ) {
+                Icon(imageVector = EblanLauncherIcons.Edit, contentDescription = null)
+            }
+
+            IconButton(
+                onClick = onResize,
+            ) {
+                Icon(imageVector = EblanLauncherIcons.Resize, contentDescription = null)
+            }
+
+            IconButton(
+                onClick = onInfo,
+            ) {
+                Icon(imageVector = EblanLauncherIcons.Info, contentDescription = null)
+            }
+
+            IconButton(
+                onClick = onDelete,
+            ) {
+                Icon(imageVector = EblanLauncherIcons.Delete, contentDescription = null)
+            }
+
+            if (!eblanAppWidgetProviderInfosByPackageName.isNullOrEmpty()) {
+                IconButton(
+                    onClick = onWidgets,
+                ) {
+                    Icon(
+                        imageVector = EblanLauncherIcons.Widgets,
+                        contentDescription = null,
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun GridItemMenu(
+    modifier: Modifier = Modifier,
+    onEdit: () -> Unit,
+    onResize: () -> Unit,
+    onDelete: () -> Unit,
+) {
+    Row(modifier = modifier) {
+        IconButton(
+            onClick = onEdit,
+        ) {
+            Icon(imageVector = EblanLauncherIcons.Edit, contentDescription = null)
+        }
+
+        IconButton(
+            onClick = onResize,
+        ) {
+            Icon(imageVector = EblanLauncherIcons.Resize, contentDescription = null)
+        }
+
+        IconButton(
+            onClick = onDelete,
+        ) {
+            Icon(imageVector = EblanLauncherIcons.Delete, contentDescription = null)
+        }
+    }
+}
+
+@Composable
+private fun WidgetGridItemMenu(
+    modifier: Modifier = Modifier,
+    showResize: Boolean,
+    onResize: () -> Unit,
+    onDelete: () -> Unit,
+) {
+    Row(modifier = modifier) {
+        if (showResize) {
+            IconButton(
+                onClick = onResize,
+            ) {
+                Icon(
+                    imageVector = EblanLauncherIcons.Resize,
+                    contentDescription = null,
+                )
+            }
+        }
+
+        IconButton(
+            onClick = onDelete,
+        ) {
+            Icon(
+                imageVector = EblanLauncherIcons.Delete,
+                contentDescription = null,
+            )
+        }
+    }
+}
