+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/component/resize/ResizeOverlay.kt
@@ -1,0 +1,107 @@
+/*
+ *
+ *   Copyright 2023 Einstein Blanco
+ *
+ *   Licensed under the GNU General Public License v3.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       https://www.gnu.org/licenses/gpl-3.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+package com.eblan.launcher.feature.home.component.resize
+
+import androidx.compose.runtime.Composable
+import com.eblan.launcher.domain.model.GridItem
+import com.eblan.launcher.domain.model.GridItemData
+import com.eblan.launcher.domain.model.GridItemSettings
+import com.eblan.launcher.domain.model.TextColor
+import com.eblan.launcher.feature.home.util.getGridItemTextColor
+import com.eblan.launcher.feature.home.util.getSystemTextColor
+
+@Composable
+internal fun ResizeOverlay(
+    gridItem: GridItem,
+    gridWidth: Int,
+    gridHeight: Int,
+    cellWidth: Int,
+    cellHeight: Int,
+    columns: Int,
+    rows: Int,
+    x: Int,
+    y: Int,
+    width: Int,
+    height: Int,
+    textColor: TextColor,
+    lockMovement: Boolean,
+    gridItemSettings: GridItemSettings,
+    onResizeGridItem: (
+        gridItem: GridItem,
+        columns: Int,
+        rows: Int,
+        lockMovement: Boolean,
+    ) -> Unit,
+) {
+    val currentTextColor = if (gridItem.override) {
+        getGridItemTextColor(
+            systemTextColor = textColor,
+            systemCustomTextColor = gridItemSettings.customTextColor,
+            gridItemTextColor = gridItem.gridItemSettings.textColor,
+            gridItemCustomTextColor = gridItem.gridItemSettings.customTextColor,
+        )
+    } else {
+        getSystemTextColor(
+            systemTextColor = textColor,
+            systemCustomTextColor = gridItemSettings.customTextColor,
+        )
+    }
+
+    when (val data = gridItem.data) {
+        is GridItemData.ApplicationInfo,
+        is GridItemData.ShortcutInfo,
+        is GridItemData.Folder,
+        is GridItemData.ShortcutConfig,
+        -> {
+            GridItemResizeOverlay(
+                gridItem = gridItem,
+                gridWidth = gridWidth,
+                gridHeight = gridHeight,
+                cellWidth = cellWidth,
+                cellHeight = cellHeight,
+                columns = columns,
+                rows = rows,
+                x = x,
+                y = y,
+                width = width,
+                height = height,
+                color = currentTextColor,
+                lockMovement = lockMovement,
+                onResizeGridItem = onResizeGridItem,
+            )
+        }
+
+        is GridItemData.Widget -> {
+            WidgetGridItemResizeOverlay(
+                gridItem = gridItem,
+                gridWidth = gridWidth,
+                gridHeight = gridHeight,
+                rows = rows,
+                columns = columns,
+                data = data,
+                x = x,
+                y = y,
+                width = width,
+                height = height,
+                color = currentTextColor,
+                lockMovement = lockMovement,
+                onResizeWidgetGridItem = onResizeGridItem,
+            )
+        }
+    }
+}
