+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/component/resize/WidgetGridItemResizeOverlay.kt
@@ -15,7 +15,7 @@
  *   limitations under the License.
  *
  */
-package com.eblan.launcher.feature.home.screen.resize
+package com.eblan.launcher.feature.home.component.resize
 
 import android.appwidget.AppWidgetManager
 import android.appwidget.AppWidgetProviderInfo
